+++ scripts/macros.js
@@ -259,3 +259,4 @@ export {improvedDuplicity, improvedDuplicityEffect} from './macros/2024/classFea
 //War Domain
 export {guidedStrike} from './macros/2024/classFeatures/cleric/war/guidedStrike.js';
 export {warPriest} from './macros/2024/classFeatures/cleric/war/warPriest.js';
+//export {warGodsBlessing} from './macros/2024/classFeatures/cleric/war/warGodsBlessing.js';