+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.3.31 Release:
+## New Generic Monster Features:
+- Activity On Rest. Note: Example "Bite" item included in the "CPR Miscellaneous" compendium.
+- Added Remove on Save and Activities config options to Reduce Max HP.
+## Bug Fixes:
+- Chat Message Dark Mode
+## API Update Notes:
+- Rest triggers now include `data` in the `trigger` object, which has a `longRest` property.
+  
 # 1.3.30 Release:
 ## New Generic Monster Features:
 - Reroll Save On Damage